feat: added header.js for UI functionality of the header display
